+++ resources/scripts/release_notes.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+cat CHANGELOG.md | awk '
+  /^## [0-9]/ {
+      release++;
+  }
+  !/^## [0-9]/ {
+      if ( release == 1 ) print;
+      if ( release > 1 ) exit;
+  }'